them account_balance cho bidder

+++ server/src/main/java/com/auction/server/model/Bidder.java
@@ -1,11 +1,12 @@
 package com.auction.server.model;
+import java.math.BigDecimal;
 import java.util.ArrayList;
 
 
 public class Bidder extends User {
 
     //Fields:
-    double account_balance;
+    BigDecimal account_balance;
     ArrayList<Auction> history_of_auction;                              //nhớ tạo lớp Auction
 
     //Constructor:
@@ -14,7 +15,7 @@ public class Bidder extends User {
                     String email, 
                     String password, 
                     String role, 
-                    double account_balance, 
+                    BigDecimal account_balance,
                     ArrayList<Auction> history_of_auction   ) {
 
         super(name, ID, email, password, role);
@@ -30,10 +31,10 @@ public class Bidder extends User {
     }
 
     // Getters and Setters:
-    public double getAccount_balance() {
+    public BigDecimal getAccount_balance() {
         return account_balance;
     }   
-    public void setAccount_balance(double account_balance) {
+    public void setAccount_balance(BigDecimal account_balance) {
         this.account_balance = account_balance;
     }
     public ArrayList<Auction> getHistory_of_auction() {
